allow the user to create custom PID namespaces
After this big refactoring the user is now able to define custom rules
to create different PID namespaces inside uprocfs.

A configuration file must be passed to the uproc daemon; the
configuration file contains all the generic rules (at the moment based
only on the UID or GID of the tasks) to associate PIDs to different
namespaces.

Example:

 uid 1000 my_username
 gid 34 backup

This will create two different namespaces in uprocfs:
  /uproc/my_username
  /uproc/backup

The first one will contain all the PIDs that run as UID 1000, the second
file will contain all the PIDs that run as GID 34.

TODO: re-read the configuration file at runtime without remounting
uprocfs.

Signed-off-by: Andrea Righi <andrea@betterlinux.com>