use a read/write locking to protect internal data structures
Use a pthread_rwlock_t instead of a pthread_spinlock_t for locking, to
increase performance of parallel readers of the filesystem and reduce
the CPU load in case of concurrent accesses to the filesystem.

Now readers of the uprocfs can all go in parallels if the writer does
not hold the lock.

Signed-off-by: Andrea Righi <andrea@betterlinux.com>