properly initialize PID hash list of a namespace
Correctly initialize all entries of the PID hash list of a namespace
using the INIT_HLIST_HEAD() macro.

Actually, there was no bug at all, since the list is allocated using
calloc(), but for a correct readability of the code it is better to
initialize all elements in the proper way.

Signed-off-by: Andrea Righi <andrea@betterlinux.com>