introduce a single reverse hash pid map
Introduce an additional hash list to map a PID to a list of namespace
elements.

This strongly speed up the retrieval of namespace elements when a task
exit, because we don't need to scan all the elements of all namespaces
anymore.

In this way we can also replace the hash list defined inside each
namespace with a plain list, reducing the memory footprint of the
application.

Signed-off-by: Andrea Righi <andrea@betterlinux.com>