small code refactoring
Do not use an unsigned long both for uid_t and gid_t, but define a union
of uid_t and gid_t and use them separately in the key structure
(uproc_key_t).

In addition, replace a strdup() with strndup() to allocate the string
that represents the name of a namespace (that is provided by the
user's configuration file).

Signed-off-by: Andrea Righi <andrea@betterlinux.com>