fix license file COPYING
For some strange reasons it seems I copied the GNU LESSER GENERAL PUBLIC
LICENSE to the COPYING file, while the comments in the source code
states that the code uses the GPL v2 license.

This is not a big issue, but fix by copying the right license to the
COPYING file (GPL v2).

Signed-off-by: Andrea Righi <andrea@betterlinux.com>